ReviewsApi/requirements.txt
asgiref==3.2.10
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
Markdown==3.2.2
mysql-connector-python==8.0.20
mysqlclient==1.4.6
pip==19.0.3
protobuf==3.12.2
pytz==2020.1
setuptools==40.8.0
six==1.15.0
sqlparse==0.3.1
wheel==0.34.2
